Crate llvm_link

llvm_link

These are bindings to the LLVM's libLTO Link Time Optimization library.

This crate currently assumes you are locally using LLVMv3.9.1 It was originally compiled on Fedora25 AMD64, so you may have trouble getting it to work on other platforms. Windows and OSX is not yet supported. To build on Fedora25 you will need to install.

sudo dnf install gcc gcc-c++ llvm-devel redhat-rpm-config ncurses-devel

To use in a Rust Project

[dependencies]
llvm_link = "0.0.1"

# Examples:

To get the local libLTO version:

use llvm_link::get_lto_version;

 println!("{}",get_lto_version());

Structs

Linker
Used to construct the Linker

ObjFile
Object Files loaded from the File System into LLVM directly

Object
Link Time Optimization Object

Enums

PIC
CodeGeneration Model

Traits

LinkerObject
Unifying properties that all linker objects posses.

Functions

get_error_msg
Get the LTO error message

get_lto_api_version
Get libLTO API version

get_lto_version
Get the LTO version